TBC721: Merge tbc721 into nft as it is like a helper
